backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.123.9",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
    "numpy>=2.3.5",
    "openai-agents>=0.6.2",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.16.1",
]
